basic scheduler and actor
