--- cgc/internal/pkg/sel1547/device.go
@@ -1,6 +1,0 @@
-package asset
-
-type Device interface {
-	ReadDeviceStatus() error
-	WriteDeviceControl() error
-}
